Added restrictions to prevent unauthorized users from deleting or editing posts that do not belong to them

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -69,6 +79,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', "Unauthorized to visit requested page.");
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -101,6 +114,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', "Unauthorized to visit requested page.");
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post removed.');
     }
